DSPBRIDGE: Compiler wanrning fix for unlocked_ioctl
Compiler wanrning fix for unlocked_ioctl

Acked-by: Omar Ramirez Luna <omar.ramirez@ti.com>

+++ drivers/dsp/bridge/rmgr/drv_interface.c
@@ -665,7 +665,7 @@ static int bridge_release(struct inode *ip, struct file *filp)
 }
 
 /* This function provides IO interface to the bridge driver. */
-static int bridge_ioctl(struct file *filp, unsigned int code,
+static long bridge_ioctl(struct file *filp, unsigned int code,
 		unsigned long args)
 {
 	int status;
